requirements.txt
asgiref==3.8.1
Django==4.2.6
django-rest-knox==5.0.2
django-rest-passwordreset==1.5.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.13.2
asgiref==3.8.1
Django==4.2.6
django-rest-knox==5.0.2
django-rest-passwordreset==1.5.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.13.2
